side bar changed to tsx
